Improvements and bugfixes to metadata writing and metadata report. The metadata report now actually writes metadata (to a temporary file). Some benefits: The report now shows what will be written to the file. The report now shows if the metadata will be deleted upon download (if $exiftool_remove_existing=true). The report now shows also metadata that will be added to the file (not just changed).
The function write_metadata() doesn't try to write the file size any more. (And the report is not displaying changes to the file size.)
Keywords are now saved separately as they should and not as a long string.
Code cleanup.

Added 5 $lang[...].
Updated the Swedish language file (complete).

Fixes issue #130.


git-svn-id: http://svn.montala.net/svn/resourcespace@4096 c08608d7-6e46-0410-86ca-f2a6f1370df5